Move utils/article_utils para data
